fix gint/gsize confusion.
Fri Dec  5 12:09:13 2003  Manish Singh  <yosh@gimp.org>

        * glib/gunidecomp.c (_g_utf8_normalize_wc): fix gint/gsize confusion.

+++ glib/gunidecomp.c
@@ -357,7 +357,7 @@ _g_utf8_normalize_wc (const gchar    *str,
 
       if (wc >= 0xac00 && wc <= 0xd7af)
         {
-          gint result_len;
+          gsize result_len;
           decompose_hangul (wc, NULL, &result_len);
           n_wc += result_len;
         }
@@ -388,7 +388,7 @@ _g_utf8_normalize_wc (const gchar    *str,
 	  
       if (wc >= 0xac00 && wc <= 0xd7af)
         {
-          gint result_len;
+          gsize result_len;
           decompose_hangul (wc, wc_buffer + n_wc, &result_len);
           n_wc += result_len;
         }
